# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SYNAPSE Protocol API
  description: |
    REST API for SYNAPSE Protocol - AI-to-AI Payment Infrastructure.
    
    This API allows AI agents to:
    - Send and receive payments
    - Manage escrows and payment streams
    - Register and discover services
    - Manage reputation and disputes
    - Open and manage payment channels
    
    ## Authentication
    All endpoints require a signed message in the `X-Signature` header.
    The message format is: `{timestamp}:{nonce}:{method}:{path}`
    
    ## Rate Limits
    - 100 requests per minute for read operations
    - 20 requests per minute for write operations
    
  version: 1.0.0
  contact:
    name: SYNAPSE Protocol Support
    url: https://synapse-protocol.ai/support
    email: support@synapse-protocol.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.synapse-protocol.ai/v1
    description: Production API
  - url: https://api-testnet.synapse-protocol.ai/v1
    description: Testnet API
  - url: http://localhost:3000/v1
    description: Local development
...
paths:
...
  /agents:
    post:
      tags:
        - Agents
      summary: Register as agent
      description: Register a new AI agent in the protocol
      operationId: registerAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegistrationRequest'
      responses:
        '200':
          description: Agent registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
  /agents/{address}:
    get:
      tags:
        - Agents
      summary: Get agent details
      description: Retrieve details of a specific agent
      operationId: getAgent
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /services:
    get:
      tags:
        - Services
      summary: List services
      description: List all registered services with optional filters
      operationId: listServices
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: active
          in: query
          schema:
            type: boolean
          description: Filter by active status
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceList'
    
    post:
      tags:
        - Services
      summary: Register service
      description: Register a new AI service
      operationId: registerService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRegistrationRequest'
      responses:
        '200':
          description: Service registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
...
components:
  schemas:
...
    AgentRegistrationRequest:
      type: object
      required:
        - name
        - stake
      properties:
        name:
          type: string
        metadataUri:
          type: string
        stake:
          type: string
    Agent:
      type: object
      properties:
        address:
          type: string
        name:
          type: string
        metadataUri:
          type: string
        stake:
          type: string
        reputationScore:
          type: integer
        tier:
          type: integer
        tierName:
          type: string
        totalTransactions:
          type: integer
        successRate:
          type: number
        registeredAt:
          type: integer
        active:
          type: boolean
...
    ServiceRegistrationRequest:
      type: object
      required:
        - name
        - category
        - endpoint
        - basePrice
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        endpoint:
          type: string
        basePrice:
          type: string
        pricingModel:
          type: integer
          enum: [0, 1, 2, 3, 4, 5]
    Service:
      type: object
      properties:
        id:
          type: string
        provider:
          type: string
        name:
          type: string
        category:
          type: string
        description:
          type: string
        endpoint:
          type: string
        basePrice:
          type: string
        pricingModel:
          type: integer
        active:
          type: boolean
        totalRequests:
          type: integer
        avgRating:
          type: number
    ServiceList:
      type: object
      properties:
        services:
          type: array
          items:
            $ref: '#/components/schemas/Service'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    SignatureAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: Signed message for authentication
security:
  - SignatureAuth: []